This is reume with bug i m trying to fix
The site is working it display experience but it only seed one out of 3 description that is only bug . I asked tucker for tutor since last 4days wish to learn how to do that. Unfortunately he is unavailable .Added Microsoft as third party did not removed google it is not working due to new api. I have authorize  done .